add basic htrace instrumentation points and configs for turning htrace on/off. defaults to off.

when merged, we should provide some sample docs on using this to solve e.g. #415 .