@cmccoy 2015-12-10
+1. Is there any non-negligible overhead with tracing disabled?