@busbey 2015-12-10
I haven't verified the overhead experimentally yet. Looking at the code for the HTrace library, when things are disabled I think it just constantly returns a singleton.