@gkamat 2015-12-11
I think it would be interesting to check out the amount of overhead HTrace imposes.  I'll run some tests with and without the change and then check performance after enabling tracing.  Could we hold off merging the change until I get some numbers?  Thanks.