@stfeng 2015-12-11
One more (minor) comment: it'd be nice if you could break this commit into two, one for the formatting changes only, and the second one contains new logic. It makes change isolation easier...

But it's up to you, I don't have a strong opinion on this (?w=0 works well for me as a reviewer ;)