@busbey 2015-12-12
AFAIK, the only formatting changes are from when I'm placing lines into a try block (and thus their indentation level should change).