@busbey 2016-01-05
rebased to current master, made sure we didn't have a System.exit within any of our htrace blocks.

@gkamat how're things looking?