@busbey 2016-02-07
Rebased to current master. incorporated feedback from @clehene.

Thanks for the review!