@saintstack 2016-04-28
I ran a basic compare of before and after applying the patch at 0f674fc7278c947acc25a596ef4a7d97ca8800c1

Client is doing ycsb workload c with 48 threads doing 100M ops.

My first test was running the client as part of a fleet of 8 nodes. In this case it the patched (but w/o htrace enabled) ran faster/did less work.

This is without the patch:
```

2016-04-28 13:26:28:617 679 sec: 10000000 operations; 889.26 current ops/sec; [CLEANUP: Count=36, Max=44191, Min=2, Avg=1236.33, 90=20, 99=44191, 99.9=44191, 99.99=44191] [READ: Count=8181, Max=634367, Min=334, Avg=3065, 90=4879, 99=16271, 99.9=209279, 99.99=592383]

Performance counter stats for '/home/stack/ycsb-0.9.0-SNAPSHOT/bin/ycsb run hbase10 -P /home/stack/ycsb-0.9.0-SNAPSHOT/workloads/workloadc -p table=ycsb -cp /home/stack/conf_hbase -p columnfamily=family -s -threads 48 -jvm-args=-Xmx8192m -Djava.security.egd=file:/dev/./urandom  -p recordcount=10000000 -p operati    oncount=10000000 -p exportfile=logs/ycsb-workloadc-measurements-ve0524-20160428T20:15:08Z.json -p exporter=com.yahoo.ycsb.measurements.exporter.JSONArrayMeasurementsExporter':
    1633454.192658 task-clock (msec)         #    2.401 CPUs utilized
        26,310,835 context-switches          #    0.016 M/sec
         4,688,282 cpu-migrations            #    0.003 M/sec
         1,265,819 page-faults               #    0.775 K/sec
 2,082,072,847,238 cycles                    #    1.275 GHz
   <not supported> stalled-cycles-frontend
   <not supported> stalled-cycles-backend
 1,071,649,307,268 instructions              #    0.51  insns per cycle
   204,398,225,103 branches                  #  125.133 M/sec
     8,028,409,082 branch-misses             #    3.93% of all branches

     680.266738512 seconds time elapsed
```

This is with the patch:

```
2016-04-28 13:40:47:499 531 sec: 10000000 operations; 1097.73 current ops/sec; [CLEANUP: Count=10, Max=37631, Min=2, Avg=3767.6, 90=14, 99=37631, 99.9=37631, 99.99=37631] [READ: Count=1439, Max=91007, Min=307, Avg=1281.14, 90=2187, 99=10471, 99.9=46239, 99.99=91007]

 Performance counter stats for '/home/stack/ycsb-0.9.0-SNAPSHOT/bin/ycsb run hbase10 -P /home/stack/ycsb-0.9.0-SNAPSHOT/workloads/workloadc -p table=ycsb -cp /home/stack/conf_hbase -p columnfamily=family -s -threads 48 -jvm-args=-Xmx8192m -Djava.security.egd=file:/dev/./urandom  -p recordcount=10000000 -p operationcount=10000000 -p exportfile=logs/ycsb-workloadc-measurements-ve0524-20160428T20:31:55Z.json -p exporter=com.yahoo.ycsb.measurements.exporter.JSONArrayMeasurementsExporter':

    1603640.139671 task-clock (msec)         #    3.012 CPUs utilized
        26,357,262 context-switches          #    0.016 M/sec
         4,693,836 cpu-migrations            #    0.003 M/sec
         1,432,836 page-faults               #    0.893 K/sec
 2,045,935,113,904 cycles                    #    1.276 GHz
   <not supported> stalled-cycles-frontend
   <not supported> stalled-cycles-backend
 1,056,620,914,643 instructions              #    0.52  insns per cycle
   199,977,636,377 branches                  #  124.702 M/sec
     7,880,502,259 branch-misses             #    3.94% of all branches

     532.391661177 seconds time elapsed
```

Let me try the patched client on its ownsome and run it a few times....