@saintstack 2016-04-28
Ok. I ran with long client running 100M ops with 48 threads before and after applying the patch. Did it three times. With the patch, we might be doing a few more instructions but about same instructions per cycle and same overall throughput roughly

Without patch
```

 2016-04-28 14:04:57,354 INFO  [Thread-2-EventThread] zookeeper.ClientCnxn: EventThread shut down
 2016-04-28 14:04:57:456 127 sec: 10000000 operations; 75348.03 current ops/sec; [READ: Count=582518, Max=35007, Min=234, Avg=596.31, 90=707, 99=7623, 99.9=14375, 99.99=20367] [CLEANUP: Count=96, Max=124671, Min=2, Avg=1317.21, 90=22, 99=785, 99.9=124671, 99.99=124671]

  Performance counter stats for '/home/stack/ycsb-0.9.0-SNAPSHOT/bin/ycsb run hbase10 -P /home/stack/ycsb-0.9.0-SNAPSHOT/workloads/workloadc -p table=ycsb -cp /home/stack/conf_hbase -p columnfamily=family -s -threads 48 -jvm-args=-Xmx8192m -Djava.security.egd=file:/dev/./urandom  -p
 recordcount=10000000 -p operati    oncount=10000000 -p exportfile=logs/ycsb-workloadc-measurements-ve0524-20160428T21:02:48Z.json -p exporter=com.yahoo.ycsb.measurements.exporter.JSONArrayMeasurementsExporter':

     1448443.518777 task-clock (msec)         #   11.264 CPUs utilized
         28,226,146 context-switches          #    0.019 M/sec
         10,228,545 cpu-migrations            #    0.007 M/sec
            729,752 page-faults               #    0.504 K/sec
  2,016,484,967,472 cycles                    #    1.392 GHz
    <not supported> stalled-cycles-frontend
    <not supported> stalled-cycles-backend
  1,003,193,720,859 instructions              #    0.50  insns per cycle
    178,052,034,558 branches                  #  122.926 M/sec
      2,544,438,220 branch-misses             #    1.43% of all branches

      128.591543660 seconds time elapsed

...
 2016-04-28 14:27:38:492 127 sec: 10000000 operations; 75753.02 current ops/sec; [READ: Count=589813, Max=36639, Min=232, Avg=601.35, 90=725, 99=8091, 99.9=12559, 99.99=18671] [CLEANUP: Count=96, Max=122879, Min=2, Avg=1296.81, 90=18, 99=770, 99.9=122879, 99.99=122879]

  Performance counter stats for '/home/stack/ycsb-0.9.0-SNAPSHOT/bin/ycsb run hbase10 -P /home/stack/ycsb-0.9.0-SNAPSHOT/workloads/workloadc -p table=ycsb -cp /home/stack/conf_hbase -p columnfamily=family -s -threads 48 -jvm-args=-Xmx8192m -Djava.security.egd=file:/dev/./urandom  -p
 recordcount=10000000 -p operati    oncount=10000000 -p exportfile=logs/ycsb-workloadc-measurements-ve0524-20160428T21:25:30Z.json -p exporter=com.yahoo.ycsb.measurements.exporter.JSONArrayMeasurementsExporter':

     1429560.096961 task-clock (msec)         #   11.119 CPUs utilized
         28,209,708 context-switches          #    0.020 M/sec
         10,189,884 cpu-migrations            #    0.007 M/sec
            570,630 page-faults               #    0.399 K/sec
  1,976,595,469,731 cycles                    #    1.383 GHz
    <not supported> stalled-cycles-frontend
    <not supported> stalled-cycles-backend
  1,000,725,835,913 instructions              #    0.51  insns per cycle
    176,393,141,848 branches                  #  123.390 M/sec
      2,517,562,830 branch-misses             #    1.43% of all branches

      128.573300989 seconds time elapsed

...
 2016-04-28 14:48:17,500 DEBUG [Thread-28] ipc.AbstractRpcClient: Stopping rpc client
 2016-04-28 14:48:17:602 130 sec: 10000000 operations; 17352.94 current ops/sec; [READ: Count=8769, Max=12607, Min=253, Avg=466.81, 90=530, 99=1250, 99.9=10519, 99.99=11831] [CLEANUP: Count=66, Max=121407, Min=2, Avg=1846.58, 90=13, 99=34, 99.9=121407, 99.99=121407]

  Performance counter stats for '/home/stack/ycsb-0.9.0-SNAPSHOT/bin/ycsb run hbase10 -P /home/stack/ycsb-0.9.0-SNAPSHOT/workloads/workloadc -p table=ycsb -cp /home/stack/conf_hbase -p columnfamily=family -s -threads 48 -jvm-args=-Xmx8192m -Djava.security.egd=file:/dev/./urandom  -p
 recordcount=10000000 -p operati    oncount=10000000 -p exportfile=logs/ycsb-workloadc-measurements-ve0524-20160428T21:46:06Z.json -p exporter=com.yahoo.ycsb.measurements.exporter.JSONArrayMeasurementsExporter':

     1462104.055059 task-clock (msec)         #   11.132 CPUs utilized
         28,187,645 context-switches          #    0.019 M/sec
         10,214,228 cpu-migrations            #    0.007 M/sec
            656,063 page-faults               #    0.449 K/sec
  2,032,209,749,245 cycles                    #    1.390 GHz
    <not supported> stalled-cycles-frontend
    <not supported> stalled-cycles-backend
  1,002,089,314,886 instructions              #    0.49  insns per cycle
    176,808,449,062 branches                  #  120.927 M/sec
      2,491,319,581 branch-misses             #    1.41% of all branches

      131.339745942 seconds time elapsed   
```

With patch
[
...
2016-04-28 13:50:37,676 INFO  [Thread-11-EventThread] zookeeper.ClientCnxn: EventThread shut down
2016-04-28 13:50:37:704 127 sec: 10000000 operations; 77356.75 current ops/sec; [READ: Count=563702, Max=24703, Min=231, Avg=585.2, 90=706, 99=7447, 99.9=12087, 99.99=19823] [CLEANUP: Count=96, Max=53023, Min=2, Avg=781.72, 90=22, 99=20527, 99.9=53023, 99.99=53023]

 Performance counter stats for '/home/stack/ycsb-0.9.0-SNAPSHOT/bin/ycsb run hbase10 -P /home/stack/ycsb-0.9.0-SNAPSHOT/workloads/workloadc -p table=ycsb -cp /home/stack/conf_hbase -p columnfamily=family -s -threads 48 -jvm-args=-Xmx8192m -Djava.security.egd=file:/dev/./urandom  -p recordcount=10000000 -p operationco
unt=10000000 -p exportfile=logs/ycsb-workloadc-measurements-ve0524-20160428T20:48:29Z.json -p exporter=com.yahoo.ycsb.measurements.exporter.JSONArrayMeasurementsExporter':

    1498744.776813 task-clock (msec)         #   11.699 CPUs utilized
        28,287,296 context-switches          #    0.019 M/sec
        10,320,958 cpu-migrations            #    0.007 M/sec
           732,257 page-faults               #    0.489 K/sec
 2,106,401,174,465 cycles                    #    1.405 GHz
   <not supported> stalled-cycles-frontend
   <not supported> stalled-cycles-backend
 1,014,689,743,992 instructions              #    0.48  insns per cycle
   179,404,941,364 branches                  #  119.703 M/sec
     2,607,635,152 branch-misses             #    1.45% of all branches

     128.111865634 seconds time elapsed


...
2016-04-28 13:55:31,779 DEBUG [Thread-49] ipc.AbstractRpcClient: Stopping rpc client
2016-04-28 13:55:31:803 125 sec: 10000000 operations; 74660 current ops/sec; [READ: Count=425694, Max=49183, Min=242, Avg=597.46, 90=718, 99=7127, 99.9=14751, 99.99=47775] [CLEANUP: Count=96, Max=45151, Min=1, Avg=487.41, 90=19, 99=864, 99.9=45151, 99.99=45151]

 Performance counter stats for '/home/stack/ycsb-0.9.0-SNAPSHOT/bin/ycsb run hbase10 -P /home/stack/ycsb-0.9.0-SNAPSHOT/workloads/workloadc -p table=ycsb -cp /home/stack/conf_hbase -p columnfamily=family -s -threads 48 -jvm-args=-Xmx8192m -Djava.security.egd=file:/dev/./urandom  -p recordcount=10000000 -p operationco
unt=10000000 -p exportfile=logs/ycsb-workloadc-measurements-ve0524-20160428T20:53:25Z.json -p exporter=com.yahoo.ycsb.measurements.exporter.JSONArrayMeasurementsExporter':

    1428719.305351 task-clock (msec)         #   11.294 CPUs utilized
        28,211,365 context-switches          #    0.020 M/sec
        10,209,394 cpu-migrations            #    0.007 M/sec
           680,489 page-faults               #    0.476 K/sec
 1,988,149,877,105 cycles                    #    1.392 GHzG
   <not supported> stalled-cycles-frontend
   <not supported> stalled-cycles-backend
 1,003,002,120,666 instructions              #    0.50  insns per cycle
   177,087,226,485 branches                  #  123.948 M/sec
     2,510,017,572 branch-misses             #    1.42% of all branches

     126.506714064 seconds time elapsed

...
2016-04-28 13:59:03,343 INFO  [Thread-11-EventThread] zookeeper.ClientCnxn: EventThread shut down
2016-04-28 13:59:03,343 DEBUG [Thread-16] ipc.AbstractRpcClient: Stopping rpc client
2016-04-28 13:59:03:377 126 sec: 10000000 operations; 76331.57 current ops/sec; [READ: Count=476309, Max=25743, Min=234, Avg=589.26, 90=700, 99=4459, 99.9=14887, 99.99=22527] [CLEANUP: Count=96, Max=58303, Min=2, Avg=624.32, 90=22, 99=783, 99.9=58303, 99.99=58303]

 Performance counter stats for '/home/stack/ycsb-0.9.0-SNAPSHOT/bin/ycsb run hbase10 -P /home/stack/ycsb-0.9.0-SNAPSHOT/workloads/workloadc -p table=ycsb -cp /home/stack/conf_hbase -p columnfamily=family -s -threads 48 -jvm-args=-Xmx8192m -Djava.security.egd=file:/dev/./urandom  -p recordcount=10000000 -p operationco
unt=10000000 -p exportfile=logs/ycsb-workloadc-measurements-ve0524-20160428T20:56:56Z.json -p exporter=com.yahoo.ycsb.measurements.exporter.JSONArrayMeasurementsExporter':

    1486900.980098 task-clock (msec)         #   11.698 CPUs utilized
        28,250,518 context-switches          #    0.019 M/sec
        10,297,318 cpu-migrations            #    0.007 M/sec
           717,429 page-faults               #    0.482 K/sec
 2,086,691,492,394 cycles                    #    1.403 GHz
   <not supported> stalled-cycles-frontend
   <not supported> stalled-cycles-backend
 1,012,490,777,213 instructions              #    0.49  insns per cycle
   178,904,362,508 branches                  #  120.320 M/sec
     2,517,984,046 branch-misses             #    1.41% of all branches

     127.102766741 seconds time elapsed](url)